+++ contao/dca/tl_c4g_firefighter_operations.php
@@ -1,5 +1,8 @@
 <?php
 
+declare(strict_types=1);
+
+use Contao\ArrayUtil;
 use Contao\CoreBundle\DataContainer\PaletteManipulator;
 
 if (!isset($GLOBALS['TL_DCA']['tl_c4g_firefighter_operations'])) {
@@ -7,7 +10,7 @@ if (!isset($GLOBALS['TL_DCA']['tl_c4g_firefighter_operations'])) {
 }
 
 // Add global operations
-array_insert($GLOBALS['TL_DCA']['tl_c4g_firefighter_operations']['list']['global_operations'], 0, [
+ArrayUtil::arrayInsert($GLOBALS['TL_DCA']['tl_c4g_firefighter_operations']['list']['global_operations'], 0, [
     'import' => [
         'label' => &$GLOBALS['TL_LANG']['tl_c4g_firefighter_operations']['import'],
         'href' => 'key=import',