+++ ecs.php
@@ -16,6 +16,7 @@ Copyright (c) $date pdir / digital agentur // pdir GmbH
 @package    content-migration-bundle
 @link       https://pdir.de
 @license    LGPL-3.0+
+@author     pdir GmbH <https://pdir.de>
 @author     Mathias Arzberger <develop@pdir.de>
 
 For the full copyright and license information, please view the LICENSE