+++ src/Controller/PageExportController.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 /*
  * Content migration bundle for Contao Open Source CMS
  *
- * Copyright (c) 2022 pdir / digital agentur // pdir GmbH
+ * Copyright (c) 2025 pdir / digital agentur // pdir GmbH
  *
  * @package    content-migration-bundle
  * @link       https://pdir.de
  * @license    LGPL-3.0+
+ * @author     pdir GmbH <https://pdir.de>
  * @author     Mathias Arzberger <develop@pdir.de>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -115,8 +116,8 @@ class PageExportController
         if (null !== $pages) {
             foreach ($pages as $page) {
                 $ids = [];
-                $ids[] = 'p'.$this->withLeadingZeroes((string)$page->pid);
-                $ids[] = 'id'.$this->withLeadingZeroes((string)$page->id);
+                $ids[] = 'p'.$this->withLeadingZeroes((string) $page->pid);
+                $ids[] = 'id'.$this->withLeadingZeroes((string) $page->id);
 
                 // write page model
                 $this->saveSerializeFile(
@@ -134,7 +135,7 @@ class PageExportController
                     foreach ($articleModel as $article) {
                         $this->saveSerializeFile(
                             $userFolder,
-                            ['id'.$this->withLeadingZeroes((string)$article->id), 'pid'.$this->withLeadingZeroes((string)$article->pid)],
+                            ['id'.$this->withLeadingZeroes((string) $article->id), 'pid'.$this->withLeadingZeroes((string) $article->pid)],
                             $article->row(),
                             'article'
                         );
@@ -146,7 +147,7 @@ class PageExportController
                             foreach ($contentModel as $content) {
                                 $this->saveSerializeFile(
                                     $userFolder,
-                                    ['pid'.$this->withLeadingZeroes((string)$article->pid), 'id'.$this->withLeadingZeroes((string)$content->id), ($content->ptable ?: 'null')],
+                                    ['pid'.$this->withLeadingZeroes((string) $article->pid), 'id'.$this->withLeadingZeroes((string) $content->id), ($content->ptable ?: 'null')],
                                     $content->row(),
                                     'content'
                                 );